+++ src/main/java/com/auction/models/Item.java
@@ -8,11 +8,11 @@ public abstract class Item extends Entity {
     private static final long serialVersionUID = 1L;
     private String itemName;
     private String description;
-    private float startingPrice;
+    private double startingPrice;
     private Seller owner;
     private List<String> imageUrls;
 
-    public Item(String itemName, String description, float startingPrice) {
+    public Item(String itemName, String description, double startingPrice) {
         super();
         this.itemName = itemName;
         this.description = description;
@@ -24,8 +24,8 @@ public abstract class Item extends Entity {
     public void setItemName(String itemName) { this.itemName = itemName; }
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
-    public float getStartingPrice() { return startingPrice; }
-    public void setStartingPrice(float startingPrice) { this.startingPrice = startingPrice; }
+    public double getStartingPrice() { return startingPrice; }
+    public void setStartingPrice(double startingPrice) { this.startingPrice = startingPrice; }
     public Seller getOwner() { return owner; }
     public void setOwner(Seller owner) { this.owner = owner; }
     public List<String> getImageUrls() { return imageUrls; }
